SimLynx

SimLynx.current_simulation

SimLynx.Allocation

SimLynx.Notice

SimLynx.Resource

SimLynx.Simulation

SimLynx.Variable

Base.get

Base.schedule

SimLynx.allocate

SimLynx.deallocate

SimLynx.event_schedule

SimLynx.interrupt

SimLynx.islessorequal

SimLynx.print_stats

SimLynx.release

SimLynx.request

SimLynx.resume

SimLynx.set!

SimLynx.start_simulation

SimLynx.suspend

SimLynx.sync!

SimLynx.update!

SimLynx.update!

SimLynx.work

SimLynx.@event

SimLynx.@process

SimLynx.@schedule

SimLynx.@thunk

SimLynx.@with_new_simulation

SimLynx.@with_resource

SimLynx.current_simulation — Constant

The current active simulation.

SimLynx.Allocation — Type

A (unit) allocation of the resource to a process.

SimLynx.Notice — Type

A notice represents the execution of an event at some (simulated) future time.

SimLynx.Resource — Type

A sharable resource with a fixed number of allocatable units. The queue maintains a list of processes waiting for the resource.

SimLynx.Simulation — Type

A simulation represents the state of an execution. Specifically, it contains the current time, current event, future event list, and control task for the simulation.

SimLynx.Variable — Type

A variable automatically maintains the history and statistics of its value over time.

Base.get — Method

Return the current value of a variable.

Base.schedule — Method

schedule(notice::Notice)
schedule(sim::Simulation, notice::Notice)

Schedule the notice on the future event list for the specified simulation, which defaults to the current simulation.

SimLynx.allocate — Method

Allocate a unit of the resource to a process.

SimLynx.deallocate — Method

Deallocate a unit of the resource from a process.

SimLynx.event_schedule — Method

event_schedule(notice::Notice, event_list::Array{Notice, 1})

Add a notice to the given event list. This routine keeps the event list in sorted (ascending) order.

SimLynx.interrupt — Method

Interrupt the execution of a waiting process.

SimLynx.islessorequal — Method

Ascending order function for event lists.

SimLynx.print_stats — Function

Print the accumulates statistics for a variable.

SimLynx.release — Method

release(resource::Resource)

Release a unit of the resource. If there are process queued for the resource, then allocate a unit of the resource to the longest waiting process. Note that this works for unit allocations.

SimLynx.request — Method

request(resource::Resource)

Request a unit of the resource. If a unit of the resource is not available, then queue the request.

SimLynx.resume — Method

Resume the execution of a suspended or interrupted process.

SimLynx.set! — Method

Set the value of a variable.

SimLynx.start_simulation — Method

start_simulation()

This is the main simulation loop.

SimLynx.suspend — Method

Suspend the execution of the current process.

SimLynx.sync! — Method

Synchronize the history and statistics for a variable. This is called before the value of a variable is changed, via set!, and before the history or statistics are used.

SimLynx.update! — Method

Accumulate running statistics over a time duration.

SimLynx.update! — Method

Tally running statistics.

SimLynx.work — Method

work(delay::Real)

Simulate the delay while work is being done. Add an event to return to this task in the future to the event list.

SimLynx.@event — Macro

@event <sig> begin
    <body>
end

Define a simulation event with the specified signature and implemented by the given body.

SimLynx.@process — Macro

@process <sig> begin
    <body>
end

Define a simulation process with the specified signature and implemented by the given body.

SimLynx.@schedule — Macro

@schedule now <expr>
@schedule at <time> <expr>
@schedule in <delta> <expr>

Schedule an event to occur in the future. The create an event and adds it to the future event list at the specified time.

SimLynx.@thunk — Macro

@thunk ex

Return a thunk (i.e., a function with no arguments) that executes the expression when called.

SimLynx.@with_new_simulation — Macro

@with_new_simulation begin
    <body>
end

Executes the body within a new simulation environment. This is the easiest way to ensure a clean simulation environment.

SimLynx.@with_resource — Macro

@with_resource resource begin
    body
end

Wrap the body in a request / release pair for the resource.